fixMake

+++ methods/srcs/join.cpp
@@ -24,7 +24,7 @@ void Join::execute(std::vector<std::string> &arg, int fd)
 		this->_channels.insert(std::make_pair(arg[1], cha));
 	}
 
-	if (this->_channels.find(arg[1])->second.get_limit() < this->_channels.find(arg[1])->second.get_fds().size() + 1)
+	if (this->_channels.find(arg[1])->second.get_limit() < (int)this->_channels.find(arg[1])->second.get_fds().size() + 1)
 	{
 		std::string msg;
 		msg.append(this->_users.find(fd)->second._prefix + "471 " + this->_channels.find(arg[1])->second.get_name() + " :Channel is already full\r\n");
